+++ test/cases/fs.c
@@ -188,7 +188,7 @@ TEST_F(fs, pread_pwrite)
     ASSERT_EQ_INT(ret, 0);
 
     ev_buf_t buf = ev_buf_make((void*)s_sample_data, strlen(s_sample_data));
-    ret = (int)ev_file_pwrite(g_test_file.file, &g_test_file.token, &buf, 1, 0,
+    ret = (int)ev_file_pwritev(g_test_file.file, &g_test_file.token, &buf, 1, 0,
         _test_file_pread_pwrite_on_write_done);
     ASSERT_EQ_INT(ret, 0);
 
@@ -196,7 +196,7 @@ TEST_F(fs, pread_pwrite)
     ASSERT_EQ_INT(ret, 0);
 
     buf = ev_buf_make(buffer, sizeof(buffer));
-    ret = (int)ev_file_pread(g_test_file.file, &g_test_file.token, &buf, 1, 0,
+    ret = (int)ev_file_preadv(g_test_file.file, &g_test_file.token, &buf, 1, 0,
         _test_file_pread_pwrite_on_read_done);
     ASSERT_EQ_INT(ret, 0);
 
@@ -259,7 +259,7 @@ TEST_F(fs, read_write)
     ASSERT_EQ_INT(ret, 0);
 
     ev_buf_t buf = ev_buf_make((void*)s_sample_data, strlen(s_sample_data));
-    ret = (int)ev_file_pwrite(g_test_file.file, &g_test_file.token, &buf, 1, 0,
+    ret = (int)ev_file_pwritev(g_test_file.file, &g_test_file.token, &buf, 1, 0,
         _test_file_read_write_on_write_done);
     ASSERT_EQ_INT(ret, 0);
 
@@ -274,7 +274,7 @@ TEST_F(fs, read_write)
     ASSERT_EQ_INT(ret, 0);
 
     buf = ev_buf_make(buffer, 4);
-    ret = (int)ev_file_read(g_test_file.file, &g_test_file.token, &buf, 1,
+    ret = (int)ev_file_readv(g_test_file.file, &g_test_file.token, &buf, 1,
         _test_file_read_write_on_read_done);
     ASSERT_EQ_INT(ret, 0);
 
@@ -284,7 +284,7 @@ TEST_F(fs, read_write)
     ASSERT_EQ_INT(ret, 0);
 
     buf = ev_buf_make(buffer, 4);
-    ret = (int)ev_file_read(g_test_file.file, &g_test_file.token, &buf, 1,
+    ret = (int)ev_file_readv(g_test_file.file, &g_test_file.token, &buf, 1,
         _test_file_read_write_on_read_done);
     ASSERT_EQ_INT(ret, 0);